+++ src/index.ts
@@ -1,2 +1,3 @@
 export * as sync from './sync';
 export * as context from './context';
+export * as channel from './channel';